refactor: added controller and service for teams
