+++ backend/uks-backend/src/main/java/com/example/uks/services/OrganisationService.java
@@ -141,65 +141,4 @@ public class OrganisationService {
                 .collect(Collectors.toList());
 
     }
-  
-  
-    public List<Team> getTeamsForMember(Integer orgId, Integer userId) {
-        Organisation organisation = organisationRepository.findByIdAndDeactivatedFalse(orgId)
-                .orElseThrow(() -> new OrganisationNotFound("Organisation not found"));
-
-        boolean isMember = organisation.getMembers().stream()
-                .anyMatch(user -> user.getId().equals(userId));
-        boolean isOwner = organisation.getOwner().getId().equals(userId);
-
-        if (!isMember && !isOwner) {
-            throw new AccessDeniedException("User is not a member of this organisation.");
-        }
-
-        return organisation.getTeams().stream().toList();
-    }
-
-
-    public Team createTeam(CreateTeamDTO dto) {
-        Organisation organisation = organisationRepository.findByIdAndDeactivatedFalse(dto.getOrganisationId())
-                .orElseThrow(() -> new OrganisationNotFound("Organisation not found"));
-
-        if (!organisation.getOwner().getId().equals(dto.getOwnerId())) {
-            throw new AccessDeniedException("Only the organisation owner can create teams.");
-        }
-
-        Team team = new Team();
-        team.setName(dto.getName());
-        team.setDescription(dto.getDescription());
-        team.setTeamPersmission(dto.getTeamPersmission());
-        team.setOrganisation(organisation);
-
-        return teamRepository.save(team);
-    }
-
-    public void addMemberToTeam(Integer teamId, Integer memberId) {
-        Team team = teamRepository.findById(teamId)
-                .orElseThrow(() -> new TeamNotFoundException("Team not found"));
-
-        User member = userRepository.findById(memberId)
-                .orElseThrow(() -> new UserNotFound("User not found"));
-
-        team.getMembers().add(member);
-        teamRepository.save(team);
-    }
-
-    public void updateTeam(Integer teamId, UpdateTeamDTO dto) {
-        Team team = teamRepository.findById(teamId)
-                .orElseThrow(() -> new TeamNotFoundException("Team not found"));
-
-        if (dto.getName() != null) {
-            team.setName(dto.getName());
-        }
-        if (dto.getDescription() != null) {
-            team.setDescription(dto.getDescription());
-        }
-
-        teamRepository.save(team);
-    }
-
-
 }
